fix: fix T-2777
