+++ packages/plugins/@nocobase/plugin-auth/src/client/authenticator.ts
@@ -1,0 +1,18 @@
+import { createContext, useContext } from 'react';
+
+export type Authenticator = {
+  name: string;
+  authType: string;
+  title?: string;
+  options?: {
+    [key: string]: any;
+  };
+  sort?: number;
+};
+
+export const AuthenticatorsContext = createContext<Authenticator[]>([]);
+
+export const useAuthenticator = (name: string) => {
+  const authenticators = useContext(AuthenticatorsContext);
+  return authenticators.find((a) => a.name === name);
+};
