+++ packages/plugins/@nocobase/plugin-auth/src/client/pages/SigninPage.tsx
@@ -1,21 +1,10 @@
 import { css } from '@emotion/css';
 import { Space, Tabs } from 'antd';
-import React, { ReactElement, createContext, createElement, useState } from 'react';
+import React, { createElement, useContext } from 'react';
 import { useTranslation } from 'react-i18next';
-import { useAPIClient, useCurrentDocumentTitle, usePlugin, useRequest, useViewport } from '@nocobase/client';
+import { useCurrentDocumentTitle, usePlugin, useViewport } from '@nocobase/client';
 import AuthPlugin, { AuthPage } from '..';
-
-export type Authenticator = {
-  name: string;
-  authType: string;
-  title?: string;
-  options?: {
-    [key: string]: any;
-  };
-  sort?: number;
-};
-
-export const AuthenticatorsContext = createContext<Authenticator[]>([]);
+import { Authenticator, AuthenticatorsContext } from '../authenticator';
 
 export const useSignInPages = (): {
   [authType: string]: AuthPage['signIn'];
@@ -58,84 +47,51 @@ export const SigninPage = () => {
   useCurrentDocumentTitle('Signin');
   useViewport();
   const signInPages = useSignInPages();
-  const api = useAPIClient();
-  const [authenticators, setAuthenticators] = useState<Authenticator[]>([]);
-  const [tabs, setTabs] = useState<
-    (Authenticator & {
-      component: ReactElement<{ authenticator: Authenticator }>;
-      tabTitle: string;
-    })[]
-  >([]);
+  const authenticators = useContext(AuthenticatorsContext);
   const customSignIn = useCustomSignIn(authenticators);
-
-  const handleAuthenticators = (authenticators: Authenticator[]) => {
-    const tabs = authenticators
-      .map((authenticator) => {
-        const page = signInPages[authenticator.authType];
-        if (!page) {
-          return;
-        }
-        return {
-          component: createElement<{
-            authenticator: Authenticator;
-          }>(page.Component, { authenticator }),
-          tabTitle: authenticator.title || page.tabTitle || authenticator.name,
-          ...authenticator,
-        };
-      })
-      .filter((i) => i);
-
-    setAuthenticators(authenticators);
-    setTabs(tabs);
-  };
-
-  const { error } = useRequest(
-    () =>
-      api
-        .resource('authenticators')
-        .publicList()
-        .then((res) => {
-          return res?.data?.data || [];
-        }),
-    {
-      onSuccess: (data) => {
-        handleAuthenticators(data);
-      },
-    },
-  );
-
-  if (error) {
-    throw error;
-  }
 
   if (!authenticators.length) {
     return <div style={{ color: '#ccc' }}>{t('Oops! No authentication methods available.')}</div>;
   }
 
+  const tabs = authenticators
+    .map((authenticator) => {
+      const page = signInPages[authenticator.authType];
+      if (!page) {
+        return;
+      }
+      return {
+        component: createElement<{
+          authenticator: Authenticator;
+        }>(page.Component, { authenticator }),
+        tabTitle: authenticator.title || page.tabTitle || authenticator.name,
+        ...authenticator,
+      };
+    })
+    .filter((i) => i);
+
   return (
-    <AuthenticatorsContext.Provider value={authenticators}>
+    <Space
+      direction="vertical"
+      className={css`
+        display: flex;
+      `}
+    >
+      {tabs.length > 1 ? (
+        <Tabs items={tabs.map((tab) => ({ label: t(tab.tabTitle), key: tab.name, children: tab.component }))} />
+      ) : tabs.length ? (
+        <div>{tabs[0].component}</div>
+      ) : (
+        <></>
+      )}
       <Space
         direction="vertical"
         className={css`
           display: flex;
         `}
       >
-        {tabs.length > 1 ? (
-          <Tabs items={tabs.map((tab) => ({ label: t(tab.tabTitle), key: tab.name, children: tab.component }))} />
-        ) : tabs.length ? (
-          <div>{tabs[0].component}</div>
-        ) : (
-          <></>
-        )}
-        <Space
-          direction="vertical"
-          className={css`
-            display: flex;
-          `}
-        >
-          {customSignIn}
-        </Space>
+        {customSignIn}
       </Space>
-    </AuthenticatorsContext.Provider>
+    </Space>
   );
 };
